requirements.txt
autopep8==1.2.2
Django==1.9.2
django-filter==0.12.0
django-rest-swagger==0.3.5
djangorestframework==3.3.2
gunicorn==19.4.5
Markdown==2.6.5
pep8==1.7.0
PyYAML==3.11
